requirements.txt
openai==1.57.0
pydantic==2.10.3
python-dotenv==1.0.1
requests==2.32.3
typing==3.7.4.3